Allow PHPUnit ^11 and ^12
Allow installing newer versions of PHPUnit.
In this library itself only PHPUnit ^10 is possible because of the PHP version constraint (PHPUnit ^11 requires at least PHP 8.2).
But this PR allows consuming packages to install a newer version of PHPUnit.

PHPUnit 12 requires dataProvider to be attributes, so all dataProvider annotations were converted to attributes.
To be able to install PHPUnit 11, psalm ^6 is required, so it was also bumped and some new issues were fixed.

Signed-off-by: Thomas Rieschl <thomas@trinet.at>